Fix ambiguous argument
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/4521


git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@27510 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/i18n/rematch.cpp
@@ -1060,7 +1060,7 @@ UBool RegexMatcher::findUsingChunk() {
 //
 //--------------------------------------------------------------------------------
 UnicodeString RegexMatcher::group(UErrorCode &status) const {
-    return group(0, status);
+    return group((int32_t)0, status);
 }
 
 UText *RegexMatcher::group(UText *dest, UErrorCode &status) const {
